main is red: two files one line over their cap
`craftsmanship` blocks every open PR, and both findings landed with recent
merges. Neither is a judgement call — the caps are the caps, and both files
were one line past them.

approvals/authority.go at 501 lines answers two questions, and only one of them
is about authority: may this seat decide this staging, and — when they do —
what else hears about it. The echo table moves to decidedecho.go, which is the
concept boundary the file already had in its section comments. A reader asking
either question was reading past the other.

mailcopy's weeklyLines at 84 splits where the retrospective does: what the week
did, what the morning list did with it, and what the message asks for once it
has finished reporting. That file's own comment says why the cap is real here
rather than arbitrary — a translator opens the section they are working on, and
a hundred lines of unrelated copy above it is noise.
